Move fast_lru_cache_test.cc into lru_cache_test.cc
