+++ CMakeLists.txt
@@ -601,6 +601,7 @@ set(SOURCES
         cache/compressed_secondary_cache.cc
         cache/fast_lru_cache.cc
         cache/lru_cache.cc
+        cache/fast_lru_cache.cc
         cache/sharded_cache.cc
         db/arena_wrapped_db_iter.cc
         db/blob/blob_fetcher.cc
@@ -1206,7 +1207,6 @@ if(WITH_TESTS)
         cache/cache_test.cc
         cache/compressed_secondary_cache_test.cc
         cache/lru_cache_test.cc
-        cache/fast_lru_cache_test.cc
         db/blob/blob_counting_iterator_test.cc
         db/blob/blob_file_addition_test.cc
         db/blob/blob_file_builder_test.cc
