+++ src.mk
@@ -408,7 +408,6 @@ TEST_MAIN_SOURCES =                                                     \
   cache/cache_test.cc                                                   \
   cache/cache_reservation_manager_test.cc                               \
   cache/lru_cache_test.cc                                               \
-  cache/fast_lru_cache_test.cc                                          \
   cache/compressed_secondary_cache_test.cc                              \
   db/blob/blob_counting_iterator_test.cc                                \
   db/blob/blob_file_addition_test.cc                                    \
